requirement.txt
amqp==5.0.6
aniso8601==9.0.1
attrs==21.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.1.0
email-validator==1.1.2
Flask==2.0.1
Flask-JWT==0.3.2
Flask-JWT-Extended==4.2.1
Flask-Login==0.5.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-RQ==0.2
Flask-WTF==0.15.1
idna==3.2
importlib-metadata==4.5.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
MongoAlchemy==0.19
mongoengine==0.23.1
packaging==20.9
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
PyJWT==2.1.0
pymongo==3.11.4
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.17.1
pytz==2021.1
redis==3.5.3
rq==1.8.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.4.1
